Move killaura code to ModKillAura.java
